Reorganized StaffAssignment source files
Jira ticket: CAMS-418
