+++ user-interface/src/App.tsx
@@ -13,7 +13,7 @@ import useFeatureFlags, { TRANSFER_ORDERS_ENABLED } from './lib/hooks/UseFeature
 import SearchScreen from './search/SearchScreen';
 import { PrivacyActFooter } from './lib/components/uswds/PrivacyActFooter';
 import { MyCasesScreen } from './my-cases/MyCasesScreen';
-import { StaffAssignmentScreen } from './staff-assignment/StaffAssignmentScreen';
+import { StaffAssignmentScreen } from './staff-assignment/screen/StaffAssignmentScreen';
 import './App.scss';
 import GlobalAlert, { GlobalAlertRef } from './lib/components/cams/GlobalAlert/GlobalAlert';
 import { UswdsAlertStyle } from './lib/components/uswds/Alert';
