+++ user-interface/src/staff-assignment/row/StaffAssignmentRow.internal.ts
