+++ user-interface/src/staff-assignment/screen/StaffAssignmentScreen.scss
