+++ user-interface/src/staff-assignment/screen/StaffAssignmentScreen.tsx
@@ -12,10 +12,10 @@ import {
 import { CamsUser } from '@common/cams/users';
 import { useRef } from 'react';
 import { SearchResults, SearchResultsRowProps } from '@/search-results/SearchResults';
-import { StaffAssignmentHeader } from './StaffAssignmentHeader';
-import { StaffAssignmentRow } from './StaffAssignmentRow/StaffAssignmentRow';
+import { StaffAssignmentHeader } from '../header/StaffAssignmentHeader';
+import { StaffAssignmentRow } from '../row/StaffAssignmentRow';
 import './StaffAssignmentScreen.scss';
-import AssignAttorneyModal, { AssignAttorneyModalRef } from './modal/AssignAttorneyModal';
+import AssignAttorneyModal, { AssignAttorneyModalRef } from '../modal/AssignAttorneyModal';
 import { CamsRole } from '@common/cams/roles';
 import { useGlobalAlert } from '@/lib/hooks/UseGlobalAlert';
